client/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==23.2.0
auto-py-to-exe==2.44.1
beautifulsoup4==4.12.3
bitarray==2.9.2
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2021.5.30
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
colorama==0.4.6
colored==2.2.4
comm==0.2.2
customtkinter==5.2.2
darkdetect==0.8.0
debugpy==1.8.2
decorator==5.1.1
deprecation==2.1.0
Eel==0.17.0
et-xmlfile==1.1.0
executing==2.0.1
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
Gooey==1.0.8.1
googletrans==3.0.0
gotrue==2.5.4
greenlet==3.0.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.6.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.0.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keyboard==0.13.5
loguru==0.7.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.0.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pdfreader==0.1.15
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.2.2
postgrest==0.16.8
prompt_toolkit==3.0.47
psutil==6.0.0
pure-eval==0.2.2
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pygtrie==2.5.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyparsing==3.1.2
pypdf==4.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==26.0.3
realtime==1.0.6
requests==2.25.1
rfc3986==1.5.0
selenium==4.22.0
setuptools==71.0.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
storage3==0.7.6
StrEnum==0.4.15
supabase==2.5.1
supafunc==0.4.6
tornado==6.4.1
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.6
wcwidth==0.2.13
webdriver-auto-update==1.2.1
webdriver-manager==4.0.1
websocket-client==1.8.0
websockets==12.0
wget==3.2
whichcraft==0.6.1
win32-setctime==1.1.0
wsproto==1.2.0
wxPython==4.2.1
zope.event==5.0
zope.interface==6.4.post2
